requirements.txt
python-telegram-bot==13.11
python-dotenv==0.19.2
tk
requests
easyocr